Consolodated alphabet
