+++ src/main/service/filterService.ts
@@ -141,4 +141,16 @@ export class FilterService {
 
     return filterValues;
   }
+
+  public static generateAlphabetObject(): object {
+    // create the object for the possible alphabet options
+    const alphabetOptions = {};
+
+    for (let i = 0; i < 26; i++) {
+      const letter = String.fromCharCode(65 + i);
+      alphabetOptions[letter] = {};
+    }
+
+    return alphabetOptions;
+  }
 }
